Start.sh stuff

+++ start.sh
@@ -278,26 +278,31 @@ fi
 arch=$(uname -s)
 # Bat binary path
 bat_path="./bin/bat/bat"
-# Get binary from link in constants file
-if [[ $arch == "Linux" ]]; then
-    bat_binary_link=$(python3 -c "from constants import BAT_LINUX_BINARY_LINK as link; print(link)")
+# Check if bat binary exists
+if ! test -f "$bat_path"; then
+    cecho -c yellow -t "bat was not found. Installing..."
+    # Get binary from link in constants file
+    if [[ $arch == "Linux" ]]; then
+        bat_binary_link=$(python3 -c "from constants import BAT_LINUX_BINARY_LINK as link; print(link)")
+    else
+        bat_binary_link=$(python3 -c "from constants import BAT_AMD_DARWIN_BINARY_LINK as link; print(link)")
+    fi
+    cecho -c yellow -t "Fetching bat binary..."
+    cd bin
+    mkdir -p bat && cd bat
+    wget $bat_binary_link
+    tar -xzf $(basename "$bat_binary_link")
+    cd $(basename "$bat_binary_link" | awk -F'.tar.gz' '{print $1}')
+    mv bat ../ && cd ../
+    chmod +x bat 
+    cecho -c green -t "Installed bat!"
+    echo "Cleaning up..."
+    find . ! -name 'bat' -type f -exec rm -f {} +
+    find . ! -name 'bat' -name '.' -name '..' -type d -exec rm -rf {} +
+    cd ../..
 else
-    bat_binary_link=$(python3 -c "from constants import BAT_AMD_DARWIN_BINARY_LINK as link; print(link)")
-
+    cecho -c green -t "Bat binary found!"
 fi
-cecho -c yellow -t "Fetching bat binary..."
-cd bin
-mkdir -p bat && cd bat
-wget $bat_binary_link
-tar -xzf $(basename "$bat_binary_link")
-cd $(basename "$bat_binary_link" | awk -F'.tar.gz' '{print $1}')
-mv bat ../ && cd ../
-chmod +x bat 
-cecho -c green -t "Installed bat!"
-echo "Cleaning up..."
-find . ! -name 'bat' -type f -exec rm -f {} +
-find . ! -name 'bat' -name '.' -name '..' -type d -exec rm -rf {} +
-cd ../..
 
 # Check if docker is installed and running
 if ! (command -v docker) >/dev/null; then
@@ -314,33 +319,6 @@ if ! (command -v docker) >/dev/null; then
             quit "Docker was not installed. Please install Docker manually to use trapp."
         fi
         cecho -c green -t "Docker installed!"
-        # Check if colima is installed
-        colima_ver="0.5.6"
-        if ! (command -v colima) >/dev/null; then
-            cecho -c yellow -t "colima was not found. Installing..."
-            # download binary
-            mkdir bin/colima && cd bin/colima
-            curl -LO https://github.com/abiosoft/colima/releases/download/v$colima_ver/colima-$(uname)-$(uname -m)
-            # Check if usr/local/bin is exists, if not create it
-            if ! test -d "/usr/local/bin"; then
-                cecho -c yellow -t "Creating /usr/local/bin..."
-                sudo mkdir /usr/local/bin
-            fi
-            # if usr/local/bin requires sudo, prompt for password
-            if [ -w "/usr/local/bin" ]; then
-                cecho -c yellow -t "Installing colima to /usr/local/bin..."
-                # install in $PATH
-                install colima-$(uname)-$(uname -m) /usr/local/bin/colima
-            else
-                cecho -c yellow -t "usr/local/bin requires sudo to install Colima. Prepare to provide sudo password..."
-                sleep 2
-                # install in $PATH
-                sudo install colima-$(uname)-$(uname -m) /usr/local/bin/colima
-            fi
-            cd ..
-        else
-            cecho -c green -t "Colima found!"
-        fi
     elif [[ $arch == "Linux" ]]; then
         # Installing docker using the convenience script
         curl -fsSL https://get.docker.com -o ./bin/get-docker.sh
@@ -362,6 +340,34 @@ if ! (command -v docker) >/dev/null; then
     else
         quit "Invalid architecture: $arch. trapp is only supported on x86_64 and arm64 versions of Darwin and Linux."
     fi
+fi
+
+# Check if colima is installed
+colima_ver="0.5.6"
+if ! (command -v colima) >/dev/null; then
+    cecho -c yellow -t "colima was not found. Installing..."
+    # download binary
+    mkdir bin/colima && cd bin/colima
+    curl -LO https://github.com/abiosoft/colima/releases/download/v$colima_ver/colima-$(uname)-$(uname -m)
+    # Check if usr/local/bin is exists, if not create it
+    if ! test -d "/usr/local/bin"; then
+        cecho -c yellow -t "Creating /usr/local/bin..."
+        sudo mkdir /usr/local/bin
+    fi
+    # if usr/local/bin requires sudo, prompt for password
+    if [ -w "/usr/local/bin" ]; then
+        cecho -c yellow -t "Installing colima to /usr/local/bin..."
+        # install in $PATH
+        install colima-$(uname)-$(uname -m) /usr/local/bin/colima
+    else
+        cecho -c yellow -t "usr/local/bin requires sudo to install Colima. Prepare to provide sudo password..."
+        sleep 2
+        # install in $PATH
+        sudo install colima-$(uname)-$(uname -m) /usr/local/bin/colima
+    fi
+    cd ..
+else
+    cecho -c green -t "Colima found!"
 fi
 
 if [[ "$(docker info 2>&1)" =~ "Cannot connect to the Docker daemon" ]]; then
